refactor: generalize approximate relative time to just approximate time
